requirements.txt
# requirements.txt
fastapi
uvicorn
tensorflow-macos  # Note: Render runs on Linux, this will be adjusted
tensorflow-metal
transformers
Pillow
numpy
scikit-learn
gunicorn          # A production-ready server